requirements.txt
# =========================
# Core API (FastAPI stack)
# =========================
fastapi~=0.114
uvicorn[standard]~=0.30
gunicorn~=21.2
orjson~=3.10
python-multipart~=0.0.9
aiofiles~=24.1

# =========================
# Settings & Config
# =========================
pydantic~=2.8
pydantic-settings~=2.4
python-dotenv~=1.0

# =========================
# Database (PostgreSQL)
# =========================
SQLAlchemy~=2.0
alembic~=1.13
psycopg2-binary~=2.9
sqlmodel==0.0.18
# =========================
# Task Queue (Celery + Redis)
# =========================
celery==5.4.0
redis~=5.0
flower~=2.0     # 可选：任务监控UI

# =========================
# HTTP Clients & Utils
# =========================
httpx[http2]~=0.27
requests~=2.32
tenacity~=8.3    # 重试装饰器/策略

# =========================
# Workflow / Templates / YAML
# =========================
PyYAML~=6.0
Jinja2~=3.1

# =========================
# Cloud & Observability
# =========================
google-cloud-storage~=2.17   # 如用 GCS 存储
prometheus-client~=0.20      # 指标上报
#LLM
google-cloud-aiplatform>=1.85.0
# （可选，作迁移动手备用）
google-genai>=0.4
google-auth==2.33.0
# =========================
# Media (optional for MVP)
# =========================
ffmpeg-python~=0.2           # 需系统安装 ffmpeg 可执行文件
pydub~=0.25                  # 简单音频处理

# =========================
# Logging
# =========================
loguru~=0.7

